Fixes voor nummertjes.

+++ Depot.DAL/Models/Tour.cs
@@ -8,7 +8,7 @@ namespace Depot.DAL.Models
     public class Tour : DbEntity
     {
         public DateTime Start { get; set; }
-        public List<int> RegisteredTickets { get; set; } = new List<int>();
+        public List<long> RegisteredTickets { get; set; } = new List<long >();
         public bool Departed { get; set; } = false;
     }
